made to work with 0.3.0
changed line width for autoformat to 120 instead of 80

+++ src/Tank.java
@@ -1,21 +1,25 @@
 public class Tank {
-	int xpos;
-	int ypos;
-	int tankAngle;
-	int turretAngle;
+	double xpos;
+	double ypos;
+	double tankAngle;
+	double turretAngle;
 
 	/**
-	 * The Tank is the main object in game. 
+	 * The Tank is the main object in game.
 	 * 
-	 * @param xpos Horizontal grid position of the Tank. 
-	 * @param ypos Vertical grid position of the Tank. 
-	 * @param tankAngle Angle at which the Tank is pointing. 
-	 * @param turretAngle Angle at which the turret of the Tank is pointing. 
+	 * @param xpos
+	 *            Horizontal grid position of the Tank.
+	 * @param ypos
+	 *            Vertical grid position of the Tank.
+	 * @param tankAngle
+	 *            Angle at which the Tank is pointing.
+	 * @param turretAngle
+	 *            Angle at which the turret of the Tank is pointing.
 	 */
-	public Tank(int xpos,int ypos, int tankAngle, int turretAngle) {
-		this.xpos=xpos;
-		this.ypos=ypos;
-		this.tankAngle=tankAngle;
-		this.turretAngle=turretAngle;
+	public Tank(double xpos, double ypos, double tankAngle, double turretAngle) {
+		this.xpos = xpos;
+		this.ypos = ypos;
+		this.tankAngle = tankAngle;
+		this.turretAngle = turretAngle;
 	}
 }
